Added list-servers command (argv[1]) that gives server list ordered by distance. (By adding number as argv[2], returns only that number of closest servers).
Dirty and bad hack (yes, yes, I should be ashamed) for latency testing.

Made it do multiple upload and download tests (increasing data ammount with each following test).
(It is still not exactly the way speedtest.net works but it is getting closer.)

Made the code a bit more messy. :)
(Will work on that when it works as I want it to work.)